requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
pyotp==2.6.0
pycryptodome==3.10.1
qrcode